third_party/blink/web_tests/fast/xsl/import-non-document-node-expected.txt

CONSOLE WARNING: crbug.com/435623334: This page uses XSLT, which being considered for removal from the web. If that happens, it is possible that this page will need to be updated to maintain functionality.
Test importing an xsl:stylesheet node as stylesheet for XSLTProcessor.

SUCCESS
